+++ CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.0-alpha.5](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.11.0-alpha.4...v1.11.0-alpha.5) (2024-11-28)
+
+### Features
+
+- change notice socket ReceiveUserOrderRecords ([5b00eb0](https://github.com/ETransferProject/etransfer-toolkit/commit/5b00eb0da95d1fede71f1f52b7140126c4a161d7))
+
+## [1.11.0-alpha.4](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.11.0-alpha.3...v1.11.0-alpha.4) (2024-11-28)
+
+### Features
+
+- **add:** add log for socket ([1ed2aaa](https://github.com/ETransferProject/etransfer-toolkit/commit/1ed2aaa877c3a69742d5077d456c116a588c2dec))
+
+## [1.11.0-alpha.3](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.11.0-alpha.2...v1.11.0-alpha.3) (2024-11-25)
+
+### Features
+
+- dev socket ([7166091](https://github.com/ETransferProject/etransfer-toolkit/commit/7166091bb36af93e07026d9f61d40c7bf4bfd37c))
+
+## [1.11.0-alpha.2](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.11.0-alpha.1...v1.11.0-alpha.2) (2024-11-25)
+
+**Note:** Version bump only for package root
+
+## [1.11.0-alpha.1](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.11.0-alpha.0...v1.11.0-alpha.1) (2024-11-25)
+
+**Note:** Version bump only for package root
+
+## [1.11.0-alpha.0](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.10.1...v1.11.0-alpha.0) (2024-11-15)
+
+### Features
+
+- change type ([3124140](https://github.com/ETransferProject/etransfer-toolkit/commit/312414072fbc3b85e21921de831519265f762e83))
+
 ## [1.10.1](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.10.0...v1.10.1) (2024-11-01)
 
 ### Features