+++ packages/request/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.11.0-alpha.5](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.11.0-alpha.4...v1.11.0-alpha.5) (2024-11-28)
+
+**Note:** Version bump only for package @etransfer/request
+
+## [1.11.0-alpha.4](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.11.0-alpha.3...v1.11.0-alpha.4) (2024-11-28)
+
+**Note:** Version bump only for package @etransfer/request
+
+## [1.11.0-alpha.3](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.11.0-alpha.2...v1.11.0-alpha.3) (2024-11-25)
+
+**Note:** Version bump only for package @etransfer/request
+
+## [1.11.0-alpha.2](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.11.0-alpha.1...v1.11.0-alpha.2) (2024-11-25)
+
+**Note:** Version bump only for package @etransfer/request
+
+## [1.11.0-alpha.1](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.11.0-alpha.0...v1.11.0-alpha.1) (2024-11-25)
+
+**Note:** Version bump only for package @etransfer/request
+
+## [1.11.0-alpha.0](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.10.1...v1.11.0-alpha.0) (2024-11-15)
+
+**Note:** Version bump only for package @etransfer/request
+
 ## [1.10.1](https://github.com/ETransferProject/etransfer-toolkit/compare/v1.10.0...v1.10.1) (2024-11-01)
 
 **Note:** Version bump only for package @etransfer/request